combination and refactoring of tools module
