+++ tools/image.py
@@ -1,6 +1,6 @@
 from PIL import Image
 
-from get_resource_path import get_resource_path
+from tools.resource import get_resource_path
 
 def png_to_ico(png_file, ico_file=None):
     if ico_file is None:
@@ -8,4 +8,5 @@ def png_to_ico(png_file, ico_file=None):
     Image.open(png_file).save(ico_file)
 
 if __name__ == "__main__":
-    png_to_ico(get_resource_path(r"..\assets\images\icon.png"))
+    # png_to_ico(get_resource_path(r"..\assets\images\icon.png"))
+    pass
