+++ tools/resource.py
